+++ src/backend/core/mda/dispatch_webhooks.py
@@ -39,7 +39,6 @@ import secrets
 import time
 import uuid as uuid_module
 from dataclasses import dataclass, field
-from datetime import datetime, timezone
 from typing import Any, Dict, List, Optional, Set, Tuple
 from urllib.parse import urlparse
 
@@ -50,6 +49,7 @@ from jmap_email import JmapEmail, parse_email
 
 from core import enums, models
 from core.mda.inbound_pipeline import Decision, InboundContext, Step
+from core.mda.webhook_payload import build_jmap_email
 from core.services.ssrf import SSRFSafeSession, SSRFValidationError
 
 from messages.celery_app import app as celery_app
@@ -444,81 +444,6 @@ def find_webhook_channels_for_mailbox(
             )
         )
     )
-
-
-# --- payload builders --- #
-
-
-def _utcdate(value: Any) -> Optional[str]:
-    """Format a datetime as a JMAP ``UTCDate`` (RFC 3339 with ``Z`` suffix).
-
-    Falls back to the raw value if it isn't a datetime — the parser may
-    have given us a pre-formatted string. ``None`` stays ``None``.
-    """
-    if isinstance(value, datetime):
-        return value.astimezone(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-    return value
-
-
-def _strip_body_part(part: Dict[str, Any]) -> Dict[str, Any]:
-    """A JMAP ``EmailBodyPart`` without our parser's project extensions.
-
-    ``content`` (raw bytes on attachment parts) and ``sha256`` are
-    parser extensions, NOT RFC 8621. Attachment bytes in particular are
-    never embedded in the JSON body — JMAP keeps them behind a
-    ``blobId`` (which we don't have at webhook-fire time), and raw bytes
-    aren't JSON-serialisable anyway. Receivers that need the bytes use
-    ``format=eml``.
-    """
-    return {k: v for k, v in part.items() if k not in ("content", "sha256")}
-
-
-def build_jmap_email(
-    parsed_email: JmapEmail, *, include_body: bool = True
-) -> Dict[str, Any]:
-    """Project the parsed JMAP ``Email`` object into the webhook payload.
-
-    ``parse_email`` already returns a strict JMAP Email object
-    (RFC 8621 §4.1), so this is mostly a copy. We stamp ``receivedAt``
-    (the moment the webhook fires) and strip the parser's project
-    extensions (``_ext`` and per-part ``content`` / ``sha256``) so the
-    body is strict JMAP. Storage-time fields (``id``, ``blobId``,
-    ``threadId``, ``mailboxIds``, ``keywords``) don't exist at
-    webhook-fire time and are simply absent.
-
-    With ``include_body=False`` the body parts, ``bodyValues`` and
-    ``attachments`` are dropped — receivers get a notification-only
-    payload (subject + envelope addresses + headers) without the
-    message body content ever leaving the instance over the wire.
-    ``hasAttachment`` is preserved so receivers can still tell whether
-    the message had any.
-    """
-    email: Dict[str, Any] = dict(parsed_email)
-    email.pop("_ext", None)  # project extension, not strict JMAP
-    email["receivedAt"] = _utcdate(datetime.now(timezone.utc))
-
-    if include_body:
-        email["textBody"] = [
-            _strip_body_part(p) for p in parsed_email.get("textBody") or []
-        ]
-        email["htmlBody"] = [
-            _strip_body_part(p) for p in parsed_email.get("htmlBody") or []
-        ]
-        email["attachments"] = [
-            _strip_body_part(p) for p in parsed_email.get("attachments") or []
-        ]
-    else:
-        for key in (
-            "textBody",
-            "htmlBody",
-            "bodyValues",
-            "bodyStructure",
-            "attachments",
-            "preview",
-        ):
-            email.pop(key, None)
-
-    return email
 
 
 # --- envelope headers --- #
